Reservation completo, falta la api de notificaciones de correo y whatsapp

+++ app/Http/Controllers/Api/CustomerController.php
@@ -73,6 +73,14 @@ class CustomerController extends Controller
     public function destroy(Customer $customer)
     {
         Gate::authorize('delete', $customer);
+        if (
+            $customer->tieneRelaciones()
+        ) {
+            return response()->json([
+                'state' => false,
+                'message' => 'No se puede eliminar este cliente porque está relacionada con otros registros.'
+            ], 400);
+        }
         $customer->delete();
         return response()->json([
             'state' => true,
